requirements.txt
Django==3.0.3
django-debug-toolbar==2.1
flake8==3.7.9
psycopg2==2.8.4
requests==2.22.0
pytest-django
pytest-cov
factory_boy
